Added Scraper Application, Models and .dockerignore
Added a new application called "scraper" and creates several models: Cinema, Film, Showtime, FilmTimes, and ScraperData. These models are designed to store data collected during the scraping process.

Film: This model represents a movie being shown in a cinema. It includes information such as the title, image, rating, description, release date, and a reference to the cinema where it is being shown.

Cinema: This model represents a cinema that screens movies. It contains information such as the name, address, and URL.

Showtime: This model represents a specific movie screening at a cinema. It includes details such as the film, cinema, and date.

FilmTimes: This model represents the time slot for a specific showtime and includes a link for reservation.

ScraperData: This model represents the data collected by the scraper. It includes information such as the cinema for which the data was collected and the scraping date.

Additionally, the .dockerignore file has been added to prevent unnecessary data from being included when creating new Docker images. This helps improve security and ensures that only relevant files are included in the Docker build process.