Modified tasks.py to Retrieve Data Using Scraper, Celery, and RabbitMQ
Changed the tasks.py file. Instead of a connection check function, we implemented a function that utilizes the scraper from scrapy.py along with Celery and RabbitMQ to retrieve data from a website and save it to the database.
We also restructured the scraper application so that links and showtime hours are not separate models but added them to the Showtime model.

We extended the Celery image with a Dockerfile.selenium file to ensure proper functionality of Selenium.
For improved performance, we added settings for Chrome in the scrapy.py file, including disabling GPU acceleration, disabling the sandbox for containers, and avoiding /dev/shm usage.