Added Data Retrieval for 10 Days in Tasks
Modified the tasks.py file to include data retrieval for 10 days from the current date, ensuring that the movie information is updated daily. This allows for continuously refreshing the data in the database.

Additionally, made the "date" field in the ScraperData model unique to prevent duplicate entries. This ensures that each entry represents a unique date of scraping.