Added City Model and Updated Tasks
Added the City model to models.py, which stores information about the city name and is linked to the CinemaInCity class through the "city" field. This allows for better organization and retrieval of data related to cinemas in specific cities.

Furthermore, made the necessary adjustments in the models within tasks.py to accommodate this change. The Cinema model now includes a foreign key reference to the City model, establishing the relationship between cinemas and cities.