Added NGINX service in docker-compose and its config
This commit introduces NGINX into our application stack to act as a reverse proxy for the Django backend. The changes are made in the `docker-compose.yml` file and a new `nginx.conf` configuration file for NGINX.

In `docker-compose.yml`:
1. A new `nginx` service has been added, using the latest official NGINX image from Docker Hub.
2. The `nginx` service is set to depend on the `backend` service, ensuring that NGINX starts only after the backend is up and running.
3. The local 'nginx' directory is mounted to '/etc/nginx/conf.d' in the container, the directory where NGINX looks for its server configuration files.
4. The web server port 80 is exposed and mapped to port 8082 on the host.

In `nginx.conf`:
1. An `upstream` block named 'backend' has been created that points to the Django server running on `web:8000`.
2. A `server` block has been set up to listen on port 80 and proxy all requests to the 'backend' upstream.
3. Various proxy headers have been configured to ensure the backend server receives correct information about the original request.

These changes allow us to control the HTTP traffic to our application more efficiently and add an extra layer of security by hiding the identity of the backend server.