Added movie img URL scraping for Helios cinema
This commit introduces a new feature for Helios cinema data scraping in scraper.py and tasks.py.

Now, along with other movie details, the URL of the movie poster is also fetched from the Helios cinema website. This information is subsequently stored in the database. The addition of movie poster URLs provides a more complete data set for movies playing in Helios cinemas.