Added movie details URL retrieval for Multikino and Helios
This commit introduces changes to both scraper.py and tasks.py files.

For both Multikino and Helios, the functionality to scrape and store the URL for movie details has been added. This information is now retrieved during the scraping process and saved into the database, enhancing the amount of information we provide about each movie.